Redefining memory management functions: Add #undef lines and a comment
Civetweb.c defines several memory management functions (mg_malloc, mg_calloc, mg_realloc, mg_free),
that should be used instead of the C-library function (malloc, calloc, realloc, free) in civetweb.c
and all *.inl files.
To ensure all programmers remember this, the original functions are redefined as "DO NOT USE".

This commit adds a corresponding comment and some #undef line to the existing redefinitions.
This is related to issue #23. Still it is intended to include civetweb.h (not civetweb.c), and it
is not recommended to redefine global symbols used in civetweb.c from outside civetweb.