Fix compile errors (original LuaXML_lib.c is not strict C)
There are many compile errors in the original LuaXML_lib.c file,
when using a MSVC C compiler. It does not define variables at the
beginning of a function, which is allowed in C++ but not in C.
This seems to be a non-standard gcc extention, which must not be
used within civetweb.
These changes might be required for all new versions of this
third party component. Maybe on the long term this library should
be replaced or rewritten.