Timeout of getreq is implemented as REQUEST_TIMEOUT in read_request
The timeout passed to getreq must not be implemented with poll,
but as a timeout in read_request. This timeout is already present
in the code, and the value is taken from the configuration,
which is accessible from the connection parameter.
Thus, no further timeout parameter is required, so it is removed.

See also issue #82