Quick fix in handle_request: callbacks that do not handle a request
This is a border case in handle_request, which is probably an abuse
of callbacks that should not be allowed. For the moment, this quick
fix ensures compatibility to previous behavior.

The problem occurs only in the following situations:
1) A request is made to an URI. For this URI, there is a callback
   registered by mg_set_request_handler.
2) The callback is called, but it returns 0 to indicate it did not
   handle the request. Since the callback did not send any data,
   the server needs to handle the request and send a reply.
3) There is a file with the same name as the request.

Previous versions did deliver the content of the file.

4) The request could have been a DELETE request.
   For a DELETE request, it does make a difference if the URL
   belongs to a static file (on the disk of the server), or the
   delete request will be forwarded to a script.
   (a) In the first case, an autorization check to change files
   on the disk of the server is mandatory.
   (b) In the second case, the autorization check is optional, and
   a different password file is used.

In step (1) above, the request was identified as a request to a
script resource, and an authorization check according to (4b)
has been performed (if required). Now in step (2), the script did
not handle the request and only in step (3) we found out that we
have a request to a real file and would have needed authorization
according to (4a).

This border case is now temporarily fixed with a goto statement
back to the authorization check. This ensures compatibility to
the previous version - if anybody used callbacks and files for
the same URI, and filters access to the file in the callback.
Probably the support for using callbacks this way should be
skipped in the future version.