Provide replacement for DEBUG_TRACE
When DEBUG_TRACE isn't used it removes the functions from the body
of the code. This can lead to subtle bugs though as the following
will result in unexpected behaviour:

    if (whatever)
      DEBUG_TRACE(trace some stuff here)
    foobar()

When DEBUG_TRACE is enabled it will be called when `whatever` is true.
However, when DEBUG_TRACE is removed the `foobar()` function call is
then absorbed into the conditional.