Make sure to set number of HTTP headers when parsing
When parsing the HTTP headers if there are no headers there is a
situation where the num_headers of the request info doesn't get set.
This leads to bugs further down the line. For example, in
handle_cgi_request the request structure is initialized by
parse_http_headers. With zero headers num_headers is uninitialized.
There is then a call to get_header which uses the num_header to loop.

This would rarely happen in practice as there is usually _always_
headers on a HTTP request. However, compilers and static analysis tools
pick up the incorrect behaviour so we should fix it up.