Make read length explicitly size_t
There is a implicit cast from an integer to size_t for the length
parameter to mg_read. This patch makes that cast implicit by casting
the len to size_t. This doesn't check if the len is negative but it
shouldn't need to because if mg_read is given a huge number to read
(because the calculation wrapped around) it will just read until the
end of the socket data.