Add explicit casts to size_t in mg_upload
These casts convert the result of signed arithmetic into the correct
size_t type that the function accept. This doesn't alter any of the
logic in the function just tells the compiler that we know what we are
doing with respect to the conversions. This patch highlights that more
thorough reviews of the overflow of the arithmetic are needed before
assuming conversion to unsigned is acceptable.