Pointer alignment is acceptable
The h_addr_list is an array of pointers to network addresses for the host
in network byte order and it terminated by a null pointer. It is safe to
cast the address to a in_addr (which is effectively a uint32_t). However
we must cast it though void* to tell the compiler that we are OK with
the 1 byte alignment of the char pointer to a uint32_t pointer.