Avoid the missing-prototypes warning
Warning option -Wmissing-prototypes raises a warning for all functions
that are neither static nor in a header. Some functions should be available
only for main.c but not in the official header (civetweb.h) since they are
subject to changes. In order to avoid this warning, additional headers
outside the include folder must be added.