A batch file for quick builds
The batch file allows for the project to be built quickly using the built
in tools of Windows to get the dependencies needed. It automatically
installs CMake if not found. If building with MinGW it installs Python and
uses the script to get MInGW.

A complete fire and forget script to build the project which might help
complete newbies.