Disabled deprecated warning
We actually have a legitimate use for GetVersion here as we would like
to know the runtime version of the Windows rather than the compile
time version.

The reason GetVersion is deprecated is because people use it to detect
the availablility of functions on newer versions of Windows. In that
case the symbols should be loaded from the kernel.dll in our case we
actually want to report the version number of Windows.