Disable clang unreachable-code warning for function ssl_id_callback
Checking sizeof(pthread_t) > sizeof(unsigned long) has to be done at compile time,
since the sizeof(...) operator can not be used in "#if" preprocessor conditions.
Thus, depending on the compiler, either the if-branch or the else-branch will be
logically dead code. There seems to be no reasonable way to prevent this, so
disable -Wunreachable-code for this "if" statement.

See #200