Don't use CMake for Lua
CMake works for civetweb.c, but it does not work for CivetWeb built with Lua.
Maybe it would be possible to create a suitable CMake build (see #94),
but running CI tests are more important.

It is still possible to build with Lua using the Makefile and the Visual Studio
solution in the current project. The Makefile and the Solution remain supported
for the Lua build.