Quick fix for #295: Sec-Wesocket-Protocol response header is wrong when there are multiple subprotocols
This issue only occurs if the Websocket constructor uses an array of protocol strings as optional "protocol" argument (see https://developer.mozilla.org/en-US/docs/Web/API/WebSockets_API/Writing_WebSocket_client_applications), and even if, many user agents do not seem to care.

This is just a quick fix: The first of the Sec-Websocket-Protocol values sent by the client is returned by the server.
However, this is not how Sec-Websocket-Protocol is meant to be used in the RFC:
In order to get the RFC intended behavior, CivetWeb would need a list of websocket subprotocols accepted by the server. Then the server must either select a subprotocol supported by client and server, or the server has to abort the hanshake if no sub-protocol is acceptable.