Remove unit test for sdup in main.c
main.c is not a part of the CivetWeb library, but only a part of the standalone executable. Which part of main.c is used is heavily depending on the system, since there are a lot of #ifdef for Windows, Linux and OSX. Windows has some graphical user interface, implemented in main.c, while Linux has not. Most parts of main.c can not be tested automatically (with the given framwork), but only manually.
Including small parts of main.c into the automatic unit test will distort test statistics.