Set server preference of ciphers.
If the server supports weak ciphers, a client can still use a weak encryption by only offering the weakest cipher a server supports. Thus, a server should only support strong ciphers. If the server only supports strong ciphers, it does not matter if the server or the client preference is used.

Still, tests for the server security seem to insist on using the server preference (see #347). There are no disadvantages in doing so. So it is set as default now.