Possible fix for #336 (3)
Note that #336 is a structural problem, caused by closing an open connection
in one thread, while a second thread is still reading from the socket.
The closing thread need to joid the reading thread before closing the socket.
This is only possible if the reading thread is not blocking on the read call
(for longer). It is also not reliable that a shutdown or close in one thread
unblocks the read call in another thread.
A possible solution is to switch to non-blocking sockets and use a poll/select
call to wait for data before calling read. This change is a deeper change and
not only affecting the issue reported in #336 (wss client), but also the http
server.
It needs some intense tests before a possible release.