Unit test: Add diagnostic message, if CGI test executable does not exist.
It would also be possible to build the test executable from a CMakeFile,
but we create it when preparing the test environment, e.g., by calling
gcc in a .travis.yml file. This way we test the build/test environment
before using the actual CMakefile or any source file.
Rarely, but nevertheless every now and then, something goes wrong when
preparing the test environment. This kind of errors are detected before
starting the CMake-build when executing .travis.yml or appveyor.yml.

If someone uses CMake on a local machine without executing the build
step for the CGI test executable manually, the test will fail.

This commit adds a diagnostic message, telling the user to create the
cgi_test.cgi executable before starting the test.
This was a remaining request from #402.

In case this kind of check is no longer required, one might still add
the build for cgi_test.cgi to the CMakeFile of the unit test.