Travis: Fix compiler versions to gcc-4.6 and clang-3.4
It seems Travis CI updated the build image some days ago.
The new image uses a newer compiler, and this causes spurious and
idiotic warnings, that cannot be deactivated using a pragma

e.g.,
  "expansion of date or time macro is not reproducible",
  -Wdate-time when using __DATE__

or

 "call to ‘__fread_chk_warn’ declared with attribute warning:
  fread called with bigger size * nmemb than length of destination buffer,
  return __fread_chk (__ptr, __bos0 (__ptr), __size, __n, __stream);"

or

  changed "'return' will never be executed" from -Wunreachable-code
  to -Wunreachable-code-return

None of them can be ignored in pragmas, so one can either disable -Werror
or just don't use this compiler version.

Since we are close to a 1.10 release, now it's not the time to deal
with hypochondriac compiler warnings.

On the long term, the compiler version should be fixed, and some
selected versions should compile warning free (-Werror), while others
should only compile error free.