Officially stop support for "-Werror" + "-Wextra/-Weverything" for GCC
Almost all lines lines in CivetWeb compile free of warnings,
even on the highest possible warning level (-Weverything).
For some lines, single warnings need to be explicitly disabled
by a #pragma, and reactivated again. These lines are manually
checked and commented, instead of disabling warnings for the
entire code.
However, for GCC disabling certain warnings does not work.
This is a well known bug in GCC since at least 2012:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53431

CivetWeb will not go to a lower warning level, because GCC
does not fix this bug. Support for -Werror for GCC is explicitly
blocked in the CMakeFile until this bug is fixed.

Focus will be to compile free of warnings with other compilers
(Visual Studio, clang).