Disable IPv6 on Travis (check for IPv6 remains on AppVeyor)
IPv6 does not work on Travis CI. There are currently many open issues
(a short list is here:
https://github.com/travis-ci/travis-ci/issues/8711#issuecomment-343658967)
but no indication if and when it will be resolved.

Disable IPv6 for the moment, so the build does not remain "failed" for weeks.
Enable it again, once the IPv6 issues are fixed by Travis.

The IPv6 checks on AppVeyor for Windows remain.
IPv6 for Linux is checked manually on an irregular base.
IPv6 for OS X (Mac OS) remains unchecked.