Fixing configuration of ports for IPv4/IPv6 dual host sockets
When checking for configuring a socket, if the port option indicates that
a dual host socket should be used (single socket to receive traffic from
IPv4 and IPv6 addresses, e.g., `+80`), the IP version returned is 10. To
signal that a socket should be used by both versions, the IPV6_ONLY socket
option should be set to 0. Prior to this commit, `set_ports_option` looked
for an IP version equal to 6 to set the option to zero, but should be
looking for a version of 10.