Add means to replace mg_cry and log_access by own implemention
As discussed in #582, in some cases it might be useful to completely
replace an internal function by some user defined one. Still, the
interfaces of these functions, and even the function names (and
therefore the #define names), may change without notice.
Please read the section "Feature selection" in the Embedded.md
manual before using it.