Unit test: Don't connect to external HTTPS server on Travis OSX test
On Travis CI worker images with OSX / MacOS, it seems to be impossible
to open an HTTPS connection to an external server. A (non encrypted)
HTTP connection to the same server works. The HTTPS connection to the
same server works from Travis CI worker images with (Ubuntu) Linux.
The connection is also possible from AppVeyor CI Windows worker images.
It seems OpenSSL 1.1 works on OSX/Travis, but 1.0 does not.