build: type mismatch warning from OPENSSL symbol
The project is configured to treat warnings as errors.  The following
warning is encountered when compiling on Visual Studio 2015 (when
generating the solution from cmake) and with OPENSSL_API_1_1 is
defined, causing the build to fail:

```
warning C4245: 'initializing': conversion from 'long' to 'unsigned
long', signed/unsigned mismatch.
```

This patch suppresses the warning by explicitly casting the symbol
from OPENSSL to the expected type.

Signed-off-by: Keith Holman <keith.holman@windriver.com>