Reuse timeout for SSL Connect and Accept operation.
The SSL_Accept and SSL_Connect operations had a timeout of 2 seconds,
for slow and bad networks, this caused disconnects in the Connect and
Accept stages. This change allow the timeout to be reused to allow
SSL_Connect and SSL_Accept to be longer.

(cherry picked from commit 1917a2222a6655a4ee7ad8368a437682c718b4ef)