Add a function "mg_get_user_context_data" + documentation
There seems to be some confusion between the `void * user_data` assigned to the
CivetWeb server context (set as parameter in the mg_start call), and the
`void * user_data` assigned to the current connection.
Add a (convenience) function to get the user data directly from the connection
object, and improve API documentation.