Add {json.lua} for a test case
Source taken from https://github.com/rxi/json.lua (MIT license)
json.lua is not a part of the CivetWeb distribution, but an independent piece of code.
A copy is used here as a part of some tests. This copy is not maintained here.
For production code, please take a recent version from the original source.
Also, if you find any problem, please create an issue at the original repository.