Modify test order in unit test
There seems to be some strange behavior on MacOS with OpenSSL1.1 on the Travis build server, that cannot be observed on another operating system or using OpenSSL1.0. Change the test order to see if connection without client certificate is possible (would be a severe bug in the OpenSSL implementation), of if it just does not return any error message (annoying, but not a security issue).
Other Systems (Linux, Windows) are not affected.