Test the test: OpenSSL / Apple / TravisCI
It seems to be possible to connect to a HTTPS server on Apple/TravisCI without any client certificate, although the server is configured to require a valid client certificate. The same configuration works on all other systems. The Travis/Apple-OpenSSL configuration requires to use an old protocol version, maybe this is related and due to an outdated test image in the CI.