Add "STOP_FLAG_NEEDS_LOCK" define
The ctx->stop_flag is a boolean flag used to stop the server. It is written
once in "mg_stop" from 0 to 1. All CivetWeb threads check this flag in various
functions in a loop.
The flag is small enough to be read and written atomically, and it is declared
volatile (C volatile, not C++ volatile), so the compiler must not assume the
value will not change in another thread.
Still, tools like thread sanitizer raise a warning for this kind of exit
notification ... there could be an compiler+architecture combination that
might not understand this pattern. I did not encounter any such combination
in more than 20 years, but ... in theory ... such a combination might exist.
Thus, the STOP_FLAG_NEEDS_LOCK preprocessor define has been added.

See also #861