Add some comments on stack sizes. Minimum recommendation: 16 kB
See #837 and #895

According to #837, some comments on stack sizes are required.
They have been added to Embedding.md

According to #895 and the linked issue in the Zephyr project,
one issue was a stack size default of 8 kB that is too low in
combination with standard buffer sizes (MG_BUF_LEN).
Set the recommended default (16 kB), so the default values
for all buffers are consistent. It is still possible to set
a lower stack size, but only if other buffer sizes are set
to reduced values as well.