Automate certificate generation script
Fully automate (Linux) shell script for generating a certificate for HTTPS.
Passwords, country, company name, ... does no longer need to be entered in this version of the script.
All information is hard coded in the script and can be modified according to your needs.
The new script also shows how to use a root certificate - when adding this to the trusted root certification
authorities (in Windows), browsers will accept a https connection to localhost without any warning.

Note: These are all just DEMO certificates for localhost. Do not use them in production.