Reduce number of tests on Travis CI
Travis CI is using a credit based system for build and test time.
Up to now, the build time was free and unlimited.
Now it is still free for open source software, but it is counting credits.

Indeed some tests that we used before are performed multiple times in
similar environments, with little added value for running the test twice.
We took some time to review the running build and test tasks and dropped
some of those tests that provide little additional test quality.