Provide sort function with CivetWeb
Using qsort_r/s caused a crash on BSD/MacOS, because of a different definition
of this function on different systems. Therefore we provide our own sorting
function. The "ShellSort" algorithm was used, because it has a small code base,
low memory requirements and is still an effective algorithm, in particular for
the typical list size we need to sort. Our application is to sort directory
entries (file and folder names, sizes, dates), so optimizing for ~100 entries
seems most reasonable.

Fixes #1048 (Tests for OSX and FreeBSD on Travis CI broken)

Fixes #200 (Travis CI FreeBSD integration)