Rewrite pattern matching (#499)
Pattern matching caused several problems. For details see #499 and
issues referenced therein. There are only minor incompatibilities
as compared to the original behavior:

Now: The ? pattern will match every character, except '\0' and '/'
Before: The ? pattern will match every character except '\0'
Reasoning: It is not useful to have ? match /.
E.g. "???" would match "abc" as well as "a/c" as well as "../".

Now: The "|x" pattern will match 0 characters.
Before: The result was -1, but this was inconsistent with "x|".
The order of an empty match pattern in a match containing "|" had
influence if the empty match returned -1 or 0.

Furthermore a pattern "**" with pattern_length of 1 lead to a crash.
This has been fixed as well.