Makefile: Ensure each install is idempotent
While it's quite alright to use install to create directories, install
can do so on the fly using the `-D` parameter. Further more, install can
make installation easier when using the `-t` flag.

The combined set of changes makes it that the installation steps are now
idempotent, and do not fail anymore due to missing directories, such as
`install-headers`for example, which does not do the right thing if the
directory does not exists.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>