Fix header generation for Lua Server Pages (Kepler syntax).
The currently used code has send_no_cache_header() and send_additional_header() but NOT actually a call to mg_response_header_send(). When NO_RESPONSE_BUFFERING is not set (default), this merely adds the additional headers to the buffer but does not send them. As a final call to mg_response_header_send() was indeed missing here, they are never sent causing, e.g., headers added through the "additional_headers" config string to be ignored for Kepler LSP. Similarly, CORS and cache-control headers were missing in this case.

Signed-off-by: DL6ER <dl6er@dl6er.de>