Tolerate socket creation error if port is optional
When users specify `[::]:80o` as port but the system has IPv6 disabled globally, e.g., with
```
GRUB_CMDLINE_LINUX="ipv6.disable=1"
```
then CivetWeb refuses to start even though it should skip the port if it cannot bind to it. However, when the port is specified as being optional, there is no point in treating non-availability of the protocol differently than the port already being taken. In this case, this port should simply be skipped.

Signed-off-by: DL6ER <dl6er@dl6er.de>